jreader: fix StringAsBytes (#22)
In the original PR with `StringAsBytes` I missed testing with easyjson.
Looking in to the implementation, there isn't a good way to get a
string-as-bytes without an allocation from `easyjson` (the easyjson
lexer _does_ have a `Bytes()` method, but its designed for decoding a
string as base64 bytes:
[link](https://github.com/mailru/easyjson/blob/v0.7.6/jlexer/lexer.go#L697)).
So while this API doesn't _help_ in the easyjson case, it isn't worse
than a caller who needs bytes doing the cast on their own, and allows
easy testing without code changes if just the build tags are flipped.

cc @cwaldren-ld
