+++ jreader/reader.go
@@ -170,35 +170,6 @@ func (r *Reader) Float64OrNull() (float64, bool) {
 	return val, true
 }
 
-// String attempts to read a string value.
-//
-// If there is a parsing error, or the next value is not a string, the return value is "" and
-// the Reader enters a failed state, which you can detect with Error(). Types other than string
-// are never converted to strings.
-func (r *Reader) String() string {
-	return string(r.StringAsBytes())
-}
-
-// StringAsBytes attempts to read a string value, returning a byte slice that indexes into the
-// original JSON bytes.  This method can be used instead of String to avoid garbage creation,
-// but care must be taken to avoid modifying the returned byte slice.
-//
-// If there is a parsing error, or the next value is not a string, the return value is nil and
-// the Reader enters a failed state, which you can detect with Error(). Types other than string
-// are never converted to strings.
-func (r *Reader) StringAsBytes() []byte {
-	r.awaitingReadValue = false
-	if r.err != nil {
-		return nil
-	}
-	val, err := r.tr.StringAsBytes()
-	if err != nil {
-		r.err = err
-		return nil
-	}
-	return val
-}
-
 // StringOrNull attempts to read either a string value or a null. In the case of a string, the
 // return values are (value, true); for a null, they are ("", false).
 //
